pointerevents/deviceproperties/pointer-event-has-device-properties-uniqueid-from-pointer-event-init.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="console"></div>

<!-- This test verifies that if the kPointerEventDeviceId flag is enabled,
     pointerEvent.deviceProperties.uniqueId can be set via PointerEventInit. If not, it is
     undefined. -->
<script>
    function CheckDeviceId(event) {
        assert_equals(event.deviceProperties.uniqueId, 1001, "uniqueId is populated");
    }

    promise_test(async () => {
        if (!window.internals)
          return;
        var deviceProps = new DeviceProperties({
          uniqueId: 1001
        });
        var downEvent = new PointerEvent("pointerdown",
          {pointerId: 1,
            bubbles: true,
            cancelable: true,
            pointerType: "pen",
            width: 100,
            height: 100,
            isPrimary: true,
            deviceProperties: deviceProps
          });
        CheckDeviceId(downEvent);
        var moveEvent = new PointerEvent("pointermove",
          {pointerId: 1,
            bubbles: true,
            cancelable: true,
            pointerType: "pen",
            width: 100,
            height: 100,
            isPrimary: true,
            deviceProperties: deviceProps
          });
        CheckDeviceId(moveEvent);
        var upEvent = new PointerEvent("pointerup",
          {pointerId: 1,
            bubbles: true,
            cancelable: true,
            pointerType: "pen",
            width: 100,
            height: 100,
            isPrimary: true,
            deviceProperties: deviceProps
          });
        CheckDeviceId(upEvent);
    }, 'PointerEvent.deviceId via PointerEventInit');
</script>
